testing 21/requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
marshmallow==3.23.2
marshmallow-sqlalchemy==1.1.0
flasgger==0.9.7.1
Flask-Limiter==3.8.0
pyodbc==4.0.39
Werkzeug==2.3.7
qrcode==7.4.2
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==8.3.4
pytest-cov==6.0.0
